relocate io validation stuff
